+++ docs/release/release_notes.md
@@ -8,8 +8,9 @@
 
 ## Added
 
-- 
+- Added an assertion for the completeness of parameters.
 
 ## Notes
 
 - `std::constant_wrapper` support is experimental as of v2.1.8.
+- The use of `std::meta::is_complete_type` is experimental as of v2.1.8.